requirements.txt
pre-commit==3.0.4
black==23.1.0
isort==5.12.0
flake8==6.0.0
pandas==1.5.3
pytest==7.2.1
apache-airflow-providers-mongo==3.0.0